refactor(functions): extract build store
