+++ src/script.js
@@ -23,7 +23,7 @@ import {
   passToParam,
   resignToParam,
   swapToParam,
-  _minimalStore,
+  buildStore,
   sortCoordinates,
 } from "./functions.js";
 
@@ -53,117 +53,6 @@ function buildDebug(query) {
     render = false;
   }
   data.render = render;
-  return data;
-}
-
-/**
- * @param {String} query
- * @returns {Store}
- * @throws {Error}
- */
-export function buildStore(query) {
-  const data = _minimalStore();
-  const urlParams = new URLSearchParams(query);
-  data.players = urlParams.getAll("ps");
-  data.version = urlParams.get("v");
-
-  /** @type {Panel[][]} */
-  const hands = [];
-  const hs = urlParams.getAll("hs");
-  for (const h of hs) {
-    hands.push(h.split("|"));
-  }
-  data.hands = hands;
-
-  /** @type {Panel[]} */
-  let jar = [];
-  const j = urlParams.get("j");
-  if (j) {
-    jar = j.split("|");
-  }
-  data.jar = jar;
-
-  const cp = urlParams.get("cp");
-  if (typeof cp === "string") {
-    data.currentPlayerId = parseInt(cp, 10);
-  } else {
-    data.currentPlayerId = 0;
-  }
-
-  const md = urlParams.get("md");
-  data.moved = md === "1";
-  const ov = urlParams.get("ov");
-  data.over = ov === "1";
-
-  data.boardMeta.width = Number(urlParams.get("bw"));
-  data.boardMeta.height = Number(urlParams.get("bh"));
-  const as = urlParams.getAll("as");
-  /** @type {Act[]} */
-  const acts = [];
-  for (const m of as) {
-    const parts = m.split("|");
-    if (parts[1] === "m") {
-      /** @type {Move} */
-      const move = {
-        type: "move",
-        playerId: parseInt(parts[0], 10),
-        coordinates: [],
-      };
-      for (let i = 0; i < parts.length - 2; i += 3) {
-        move.coordinates.push({
-          x: parseInt(parts[i + 2], 10),
-          y: parseInt(parts[i + 3], 10),
-          panel: parts[i + 4],
-        });
-      }
-      move.coordinates = sortCoordinates(move.coordinates);
-      acts.push(move);
-    } else if (parts[1] === "p") {
-      /** @type {Pass} */
-      const pass = {
-        type: "pass",
-        playerId: parseInt(parts[0], 10),
-      };
-      acts.push(pass);
-    } else if (parts[1] === "r") {
-      /** @type {Resign} */
-      const resign = {
-        type: "resign",
-        playerId: parseInt(parts[0], 10),
-      };
-      acts.push(resign);
-    } else if (parts[1] === "s") {
-      /** @type {Swap} */
-      const swap = {
-        type: "swap",
-        playerId: parseInt(parts[0], 10),
-        panels: parts.slice(2),
-      };
-      acts.push(swap);
-    }
-  }
-  data.acts = acts;
-
-  /** @type {BoardPanel[][]} */
-  const board = [];
-  for (let i = 0; i < data.boardMeta.height; i++) {
-    board.push(new Array(data.boardMeta.width).fill(null));
-  }
-  data.board = board;
-
-  for (const a of data.acts) {
-    if (a.type === "move") {
-      for (const c of /** @type {Move} */ (a).coordinates) {
-        if (
-          typeof data.board[c.y] === "undefined" ||
-          typeof data.board[c.y][c.x] === "undefined"
-        ) {
-          throw new Error(`board x: ${c.x}, y: ${c.y} is unavailable`);
-        }
-        data.board[c.y][c.x] = c.panel;
-      }
-    }
-  }
   return data;
 }
